% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lookup.R
\docType{data}
\name{lookup}
\alias{lookup}
\title{Character to Morse Code Lookup}
\format{
A named character vector of length 36.
}
\source{
Internatinoal Morse Code.
}
\usage{
lookup
}
\description{
A mapping of Morse Code signals to letters (A to Z) and digits (0 to 9).
}
\keyword{datasets}
