% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{morse_lookup}
\alias{morse_lookup}
\title{Character to Morse Code Lookup}
\format{
A named character vector of length 52.
}
\source{
International Morse Code.
}
\usage{
morse_lookup
}
\description{
A mapping of Morse Code signals to letters, digits and punctuation.
}
\keyword{datasets}
