% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remorse-package.R
\docType{package}
\name{remorse-package}
\alias{remorse}
\alias{remorse-package}
\title{remorse: Translate Text to Morse to Sound}
\description{
Converts a string input to Morse Code and back. Plays the Morse Code as audible dits and dahs (dots and dashes).
}
\author{
\strong{Maintainer}: Matt Dray \email{mwdray@gmail.com}

}
\keyword{internal}
