% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/translate.R
\name{txt2morse}
\alias{txt2morse}
\title{Convert a String to Morse Code}
\usage{
txt2morse(string)
}
\arguments{
\item{string}{Character.}
}
\value{
A character string.
}
\description{
Provide a string of text and it'll be converted to the corresponding Morse
Code.
}
\details{
Only certain characters can be converted to and from Morse Code.
See the inbuilt \code{\link{lookup}} dataset.
}
\examples{
txt2morse("hello there")
}
