% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/translate.R
\name{morse2sfx}
\alias{morse2sfx}
\title{Play Morse Code as Sound}
\usage{
morse2sfx(string, dit_length = 0.05, play = TRUE)
}
\arguments{
\item{string}{Character. Some Morse Code expressed as 'dits' ('.')
'dahs' ('-') and spaces (' '). Ideally produced by function
\code{\link{txt2morse}}. See details.}

\item{dit_length}{Numeric. Length of a 'dit' in seconds. The length of a
'dah' will be three times the \code{dit_length}.}

\item{play}{Logical. Should sounds be played? Defaults to \code{TRUE}.
Included for testing purposes.}
}
\value{
Nothing. Sounds will play if \code{play = TRUE}.
}
\description{
Provide a Morse Code string and it will be played over your computer's
speakers.
}
\details{
Only certain characters can be converted to and from Morse Code.
See the inbuilt \code{\link{lookup}} dataset.
}
\examples{
\dontrun{txt2morse("hello there") |> morse2txt() |> morse2sfx}
}
